economy_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic>=1.10,<2.0
httpx
pandas
sdmx1
msal
pyyaml
pytest
scikit-learn
statsmodels
networkx
websockets
